# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Sicherheitsdienst-Tool API
  version: 1.0.0
  description: API v1 für Auth, Mitarbeiter, Sites, Schichten, Zuweisungen, Zeiterfassung, Vorfälle, Benachrichtigungen.
servers:
  - url: http://localhost:3000/api/v1
  - url: http://localhost:3001/api/v1
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        errors:
          type: array
          items:
            type: object
            additionalProperties: true
      required: [success, code, message]
    ValidationError:
      type: object
      properties:
        message:
          type: string
          example: "Eingaben sind ungültig."
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: "email"
              message:
                type: string
                example: "Muss eine gültige E-Mail sein."
              code:
                type: string
                nullable: true
                example: "invalid_email"
      required: [message]
...
    Site:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        address:
          type: string
        city:
          type: string
        postalCode:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, address, city, postalCode, createdAt, updatedAt]
    SiteCreate:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        city:
          type: string
        postalCode:
          type: string
      required: [name, address, city, postalCode]
    SiteUpdate:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        city:
          type: string
        postalCode:
          type: string
...
  responses:
    BadRequest:
      description: Ungültige Anfrage
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            code: BAD_REQUEST
            message: Anfrageparameter sind ungültig.
    Unauthorized:
      description: Nicht autorisiert
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            code: UNAUTHORIZED
            message: Token fehlt oder ist ungültig.
    Forbidden:
      description: Verboten
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            code: FORBIDDEN
            message: Keine Berechtigung.
    NotFound:
      description: Nicht gefunden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            code: NOT_FOUND
            message: Ressource nicht gefunden.
    Conflict:
      description: Konflikt
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            code: CONFLICT
            message: Ressourcenkonflikt.
    ValidationError:
      description: Validierungsfehler
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
          example:
            message: Eingaben sind ungültig.
            errors:
              - field: email
                message: Muss eine gültige E-Mail sein.
                code: invalid_email
    TooManyRequests:
      description: Zu viele Anfragen
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            code: TOO_MANY_REQUESTS
            message: Rate-Limit erreicht.
    InternalServerError:
      description: Serverfehler
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            code: INTERNAL_SERVER_ERROR
            message: Unerwarteter Fehler.
    ServiceUnavailable:
      description: Dienst nicht verfügbar
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            code: SERVICE_UNAVAILABLE
            message: Wartung oder temporärer Ausfall.
...
paths:
...
  /sites:
    get:
      operationId: listSites
      summary: Sites auflisten
      description: 'CSV/XLSX via Accept-Header (siehe /employees)'
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          description: Seite (beginnend bei 1)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          description: Elemente pro Seite (max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: sortBy
          in: query
          description: Sortierfeld
          schema:
            type: string
            enum: [name, city, postalCode, createdAt, updatedAt]
        - name: sortDir
          in: query
          description: Sortierrichtung
          schema:
            type: string
            enum: [asc, desc]
            default: asc
        - name: filter[name]
          in: query
          description: Filtert Sites nach Name (contains)
          schema:
            type: string
        - name: filter[city]
          in: query
          description: Filtert Sites nach Stadt (contains)
          schema:
            type: string
        - name: filter[postalCode]
          in: query
          description: Filtert Sites nach Postleitzahl (contains)
          schema:
            type: string
      responses:
        '200':
          description: Liste der Sites (paginierbar)
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Site'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      pageSize:
                        type: integer
                      total:
                        type: integer
                      totalPages:
                        type: integer
                  sort:
                    type: object
                    properties:
                      by:
                        type: string
                      dir:
                        type: string
                        enum: [asc, desc]
                  filters:
                    type: object
              example:
                data:
                  - id: "7c1f2e94-5678-4a12-9cde-abcdef123456"
                    name: "Messe Berlin"
                    address: "Messedamm 22"
                    city: "Berlin"
                    postalCode: "14055"
                    createdAt: "2024-08-01T12:00:00Z"
                    updatedAt: "2024-08-15T12:00:00Z"
                pagination:
                  page: 1
                  pageSize: 20
                  total: 1
                  totalPages: 1
                sort:
                  by: name
                  dir: asc
                filters:
                  city: Berlin
            text/csv:
              schema:
                type: string
                format: binary
              example: |
                id,name,address,city,postalCode,createdAt,updatedAt
                7c1f2e94-5678-4a12-9cde-abcdef123456,Messe Berlin,Messedamm 22,Berlin,14055,2024-08-01T12:00:00Z,2024-08-15T12:00:00Z
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
              example: (xlsx)
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    post:
      operationId: createSite
      summary: Site anlegen
      description: 'RBAC: ADMIN oder DISPATCHER.'
      x-required-roles:
        - ADMIN
        - DISPATCHER
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SiteCreate'
            example:
              name: "Messe Berlin"
              address: "Messedamm 22"
              city: "Berlin"
              postalCode: "14055"
      responses:
        '201':
          description: Site erstellt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Site'
              example:
                id: "7c1f2e94-5678-4a12-9cde-abcdef123456"
                name: "Messe Berlin"
                address: "Messedamm 22"
                city: "Berlin"
                postalCode: "14055"
                createdAt: "2024-08-01T12:00:00Z"
                updatedAt: "2024-08-15T12:00:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /sites/{id}:
    get:
      operationId: getSite
      summary: Site abrufen
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Site
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Site'
              example:
                id: "7c1f2e94-5678-4a12-9cde-abcdef123456"
                name: "Messe Berlin"
                address: "Messedamm 22"
                city: "Berlin"
                postalCode: "14055"
                createdAt: "2024-08-01T12:00:00Z"
                updatedAt: "2024-08-15T12:00:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    put:
      operationId: updateSite
      summary: Site aktualisieren
      description: 'RBAC: ADMIN oder DISPATCHER.'
      x-required-roles:
        - ADMIN
        - DISPATCHER
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SiteUpdate'
            example:
              name: "Messe Berlin (Nord)"
              address: "Messedamm 24"
              city: "Berlin"
              postalCode: "14055"
      responses:
        '200':
          description: Aktualisiert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Site'
              example:
                id: "7c1f2e94-5678-4a12-9cde-abcdef123456"
                name: "Messe Berlin (Nord)"
                address: "Messedamm 24"
                city: "Berlin"
                postalCode: "14055"
                createdAt: "2024-08-01T12:00:00Z"
                updatedAt: "2024-09-01T10:00:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    delete:
      operationId: deleteSite
      summary: Site löschen
      description: 'RBAC: ADMIN.'
      x-required-roles:
        - ADMIN
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Gelöscht
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...